Seperated chatbot and pointsbot into seperate files, Made welcoming more general
